requirements.txt
# Core dependencies
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1

# Database
mysqlclient==2.2.4
djongo==1.3.6
pymongo==4.6.1

# Authentication & Security
PyJWT==2.8.0
django-filter==24.1

# Task Queue
celery==5.3.6
redis==5.0.1

# API Documentation
drf-yasg==1.21.7

# Testing
pytest==8.0.2
pytest-django==4.8.0

# Development
black==24.2.0
flake8==7.0.0
isort==5.13.2

# Monitoring & Logging
sentry-sdk==1.40.4
python-json-logger==2.0.7

# Email & Notifications
sendgrid==6.11.0
twilio==8.12.0

# File Storage
boto3==1.34.34  # For AWS S3
minio==7.2.5    # For MinIO

# Payment Integration
stripe==7.11.0 